move routes into routes.py and initialise FLask app in __init__.py
